move sponsor to related to serie
