+++ app/Application.php
@@ -118,7 +118,6 @@ class Application extends LaravelApplication
     {
         $models = [
             Submission::class,
-            Sponsor::class,
             Topic::class,
             NavigationMenu::class,
             Block::class,
@@ -142,6 +141,7 @@ class Application extends LaravelApplication
             Timeline::class,
             CommitteeRole::class,
             SpeakerRole::class,
+            Sponsor::class,
         ];
 
         foreach ($models as $model) {
