--- app/Panel/Administration/Livewire/SponsorSetting.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Panel\Administration\Livewire;
-
-use App\Panel\Conference\Livewire\Forms\Conferences\SponsorSetting as BaseSponsorSetting;
-
-
-class SponsorSetting extends BaseSponsorSetting
-{
-
-}
